tampil pesanan & detail pesanan

+++ Website/bostani_web/database/migrations/2023_04_29_043802_create_t_item_pesanan_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->foreignId('shop_item_id')->nullable();
             $table->integer('item_purchase_price');
             $table->integer('item_selling_price');
-            $table->integer('item_size');
+            $table->decimal('item_size');
         });
     }
